🐛 (ExpansionPanels): unexpected IsActive paramter; ActiveClass is not working
