+++ src/Masa.Blazor/Components/ExpansionPanels/MExpansionPanel.razor
@@ -1,5 +1,5 @@
 ﻿@namespace Masa.Blazor
-@inherits MGroupItem<MExpansionPanels>
+@inherits Masa.Blazor.Components.ItemGroup.MGroupableBase<MExpansionPanels>
 
 <CascadingValue Value="this" IsFixed>
     <div class="@GetClass()"
