+++ src/Masa.Blazor/Components/ExpansionPanels/MExpansionPanel.razor.cs
@@ -1,6 +1,8 @@
-﻿namespace Masa.Blazor;
+﻿using Masa.Blazor.Components.ItemGroup;
 
-public partial class MExpansionPanel : MGroupItem<MExpansionPanels>
+namespace Masa.Blazor;
+
+public partial class MExpansionPanel : MGroupableBase<MExpansionPanels>
 {
     public MExpansionPanel() : base(GroupType.ExpansionPanels, true)
     {
@@ -15,7 +17,10 @@ public partial class MExpansionPanel : MGroupItem<MExpansionPanels>
     [Parameter]
     [MasaApiParameter(ReleasedOn = "v1.6.0")]
     public string? Text { get; set; }
-
+    
+    [Parameter]
+    public RenderFragment? ChildContent { get; set; }
+    
     public bool NextActive => ItemGroup != null && ItemGroup.NextActiveKeys.Contains(Value);
 
     public bool IsDisabled => ItemGroup != null && (ItemGroup.Disabled || Disabled);
@@ -31,13 +36,6 @@ public partial class MExpansionPanel : MGroupItem<MExpansionPanels>
 
     private static Block _block = new("m-expansion-panel");
     private ModifierBuilder _modifierBuilder = _block.CreateModifierBuilder();
-
-    protected override void OnParametersSet()
-    {
-        base.OnParametersSet();
-
-        ActiveClass ??= "m-item--active";
-    }
 
     protected override IEnumerable<string> BuildComponentClass()
     {
