+++ src/Masa.Blazor/Components/ItemGroup/MGroupItem.cs
@@ -1,5 +1,4 @@
 ﻿using Masa.Blazor.Components.ItemGroup;
-using Masa.Blazor.Mixins;
 
 namespace Masa.Blazor;
